spill disk_cache/disk_cache_key
