+++ src/query/storages/common/cache/src/providers/disk_cache/disk_cache.rs
@@ -12,10 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::fmt::Display;
 use std::fs;
 use std::fs::File;
-use std::hash::Hasher;
 use std::io::IoSlice;
 use std::io::Read;
 use std::io::Write;
@@ -41,46 +39,13 @@ use log::warn;
 use parking_lot::RwLock;
 use rayon::prelude::*;
 use rayon::ThreadPoolBuilder;
-use siphasher::sip128;
-use siphasher::sip128::Hasher128;
 
 use crate::CacheAccessor;
+use crate::DiskCacheKey;
 
 pub struct DiskCache<C> {
     cache: C,
     root: PathBuf,
-}
-
-#[derive(Debug)]
-pub struct DiskCacheKey(String);
-
-impl<S> From<S> for DiskCacheKey
-where S: AsRef<str>
-{
-    // convert key string into hex string of SipHash 2-4 128 bit
-    fn from(key: S) -> Self {
-        let mut sip = sip128::SipHasher24::new();
-        let key = key.as_ref();
-        sip.write(key.as_bytes());
-        let hash = sip.finish128();
-        let hex_hash = hex::encode(hash.as_bytes());
-        DiskCacheKey(hex_hash)
-    }
-}
-
-impl From<&DiskCacheKey> for PathBuf {
-    fn from(cache_key: &DiskCacheKey) -> Self {
-        let prefix = &cache_key.0[0..3];
-        let mut path_buf = PathBuf::from(prefix);
-        path_buf.push(Path::new(&cache_key.0));
-        path_buf
-    }
-}
-
-impl Display for DiskCacheKey {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0.clone())
-    }
 }
 
 impl<C> DiskCache<C>
